+++ Tests/TypeFactoryTest.php
@@ -212,7 +212,7 @@ class TypeFactoryTest extends TestCase
         $this->assertEquals(new ArrayShapeType(['foo' => ['type' => Type::bool(), 'optional' => false]]), Type::arrayShape(['foo' => Type::bool()]));
         $this->assertEquals(new ArrayShapeType(
             shape: ['foo' => ['type' => Type::bool(), 'optional' => false]],
-            extraKeyType: Type::union(Type::int(), Type::string()),
+            extraKeyType: Type::arrayKey(),
             extraValueType: Type::mixed(),
         ), Type::arrayShape(['foo' => Type::bool()], sealed: false));
         $this->assertEquals(new ArrayShapeType(
@@ -220,6 +220,11 @@ class TypeFactoryTest extends TestCase
             extraKeyType: Type::string(),
             extraValueType: Type::bool(),
         ), Type::arrayShape(['foo' => Type::bool()], extraKeyType: Type::string(), extraValueType: Type::bool()));
+    }
+
+    public function testCreateArrayKey()
+    {
+        $this->assertEquals(new UnionType(Type::int(), Type::string()), Type::arrayKey());
     }
 
     /**
@@ -275,7 +280,7 @@ class TypeFactoryTest extends TestCase
         yield [Type::dict(Type::bool()), ['a' => true, 'b' => false]];
         yield [Type::array(Type::string()), [1 => 'foo', 'bar' => 'baz']];
         yield [Type::array(Type::nullable(Type::bool()), Type::int()), [1 => true, 2 => null, 3 => false]];
-        yield [Type::collection(Type::object(\ArrayIterator::class), Type::mixed(), Type::union(Type::int(), Type::string())), new \ArrayIterator()];
+        yield [Type::collection(Type::object(\ArrayIterator::class), Type::mixed(), Type::arrayKey()), new \ArrayIterator()];
         yield [Type::collection(Type::object(\Generator::class), Type::string(), Type::int()), (fn (): iterable => yield 'string')()];
         yield [Type::collection(Type::object($arrayAccess::class)), $arrayAccess];
     }